+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/home/SecondaryArchiveRepository.kt
@@ -52,6 +52,9 @@ class SecondaryArchiveRepository(
 
         val texts: OptionalLoadable<String> =
             combineTexts(
+                OptionalLoadable.LoadedAvailable(
+                    if (!connectionStatus.isConnected) listOf("Disconnected") else emptyList(),
+                ),
                 addTexts,
                 syncTexts,
             ).mapLoadedData { it.joinToString(", ") }