+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/home/components/SecondaryArchiveRepositoryRow.kt
@@ -4,6 +4,8 @@ import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.unit.dp
 import compose.icons.TablerIcons
@@ -29,6 +31,7 @@ fun SecondaryArchiveRepositoryRow(
     SectionCardBottomListItem(
         title = name,
         statusText = nonPrimaryRepository.texts,
+        modifier = if (!nonPrimaryRepository.connectionStatus.isConnected) Modifier.alpha(0.6f) else Modifier,
         icon = {
             if (nonPrimaryRepository.isLoading) {
                 CircularProgressIndicator(